// src/cyfs-stack-test-typescript/cyfs_node/cyfs_node.d.ts
    interface Uint8Array {
        offset(offset: number): Uint8Array;
        offsetUint16Array(offset: number): Uint16Array;
        offsetUint32Array(offset: number): Uint32Array;
        offsetView(offset: number): DataView;
        toHex(): string;
        fromHex(hex: string): BuckyResult<Uint8Array>;
    }
// ...
export enum Protocol {
    Unk = 0,
    Tcp = 1,
    Udp = 2
}
// ...
export class SocketAddr implements RawEncode {
    is_ipv4: boolean;
    ip: string;
    port: number;
    constructor(is_ipv4: boolean, ip: string, port: number);
    toString(): string;
    static fromString(is_ipv4: boolean, addr: string): BuckyResult<SocketAddr>;
    raw_measure(ctx?: any): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
export class Endpoint implements RawEncode {
    is_static_wan: boolean;
    protocol: Protocol;
    addr: SocketAddr;
    constructor(is_static_wan: boolean, protocol: Protocol, addr: SocketAddr);
    static fromString(endpoint: string): BuckyResult<Endpoint>;
    toString(): string;
    flag(): number;
    raw_measure(ctx?: any): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
// ...
export type Option<T> = (SomeOption<T> | NoneOption) & BaseOption<T>;
export function Some<T>(val: T): Option<T>;
export const None: NoneOption;
// ...
export enum RawEncodePurpose {
        Serialize = 0,
        Hash = 1
}
export interface RawEncode {
        raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf?: (purpose?: RawEncodePurpose) => BuckyResult<Uint8Array>;
        raw_hash_encode?: () => BuckyResult<HashValue>;
}
// ...
export interface Compareable<T> {
        hashCode(): symbol;
        equals(other: T): boolean;
}
// ...
export type BuckyResult<T> = Result<T, BuckyError>;
// ...
export class HashValue implements RawEncode {
    m_buf: Uint8Array;
    constructor(buf: Uint8Array);
    as_slice(): Uint8Array;
    length(): number;
    static default(): HashValue;
    static copy_from_slice(buf: Uint8Array): HashValue;
    static hash_data(data: Uint8Array): HashValue;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    to_base_58(): string;
    to_hex_string(): string;
    static from_hex_string(s: string): BuckyResult<HashValue>;
}
export class HashValueDecoder implements RawDecode<HashValue> {
    raw_decode(buf: Uint8Array): BuckyResult<[HashValue, Uint8Array]>;
}

export abstract class PrivateKey implements RawEncode {
    readonly type: number;
    constructor(type: number);
    abstract raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
    abstract raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
    toString(): string;
    static generate_rsa(bits: number): BuckyResult<PrivateKey>;
    static generate_secp256k1(): BuckyResult<PrivateKey>;
    static generate_sm2(): BuckyResult<PrivateKey>;
    abstract public(): PublicKey;
    abstract sign(data: Uint8Array, sign_source: SignatureSource): Signature;
    abstract decrypt(input: Uint8Array, out: Uint8Array): BuckyResult<number>;
    to_vec(): BuckyResult<Uint8Array>;
}
// ...
export type SignatureSource = SignatureRefIndex | SignatureObject | SignatureKey;
export type SignatureSourceLite = number | ObjectLink | PublicKeyValue;
export class Signature implements RawEncode {
        m_sign_source: SignatureSource;
        m_sign_key_index: number;
        m_sign_time: JSBI;
        m_sign: SignData;
        constructor(sign_source: SignatureSource, sign_key_index: number, sign_time: JSBI, sign: SignData);
        get sign_key_index(): number;
        get sign_time(): JSBI;
        get sign_source(): SignatureSource;
        get sign(): SignData;
        sign_source_with_ref_index(): number;
        raw_measure(): BuckyResult<number>;
        raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
// ...
export type PublicKey = RSAPublicKey | Secp256k1PublicKey | SM2PublicKey;
// ...
export class Area implements RawEncode {
    m_country: number;
    m_carrier: number;
    m_city: number;
    m_inner: number;
    constructor(country: number, carrier: number, city: number, inner: number);
    get country(): number;
    get carrier(): number;
    get city(): number;
    get inner(): number;
    static default(): Area;
    static from_str(s: string): BuckyResult<Area>;
    clone(): Area;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    get_area_info(locale?: string): {
        country_name: string;
        state_name: string;
        city_name: string;
    } | null;
    toString(): string;
}
// ...
export enum DeviceCategory {
    OOD = 0,
    Server = 1,
    PC = 2,
    Router = 3,
    AndroidMobile = 4,
    AndroidPad = 5,
    AndroidWatch = 6,
    AndroidTV = 7,
    IOSMobile = 8,
    IOSPad = 9,
    IOSWatch = 10,
    SmartSpeakers = 11,
    Browser = 12,
    IoT = 13,
    SmartHome = 14,
    VirtualOOD = 15,
    Unknown = 255
}
export function number_2_devicecategory(x: number): DeviceCategory;
export class DeviceDescTypeInfo extends DescTypeInfo {
    obj_type(): number;
    sub_desc_type(): SubDescType;
}
export class DeviceDescContent extends DescContent {
    constructor(unique_id: UniqueId);
    type_info(): DescTypeInfo;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    unique_id(): UniqueId;
}
export class DeviceDescContentDecoder extends DescContentDecoder<DeviceDescContent> {
    type_info(): DescTypeInfo;
    raw_decode(buf: Uint8Array): BuckyResult<[DeviceDescContent, Uint8Array]>;
}
export class DeviceBodyContent extends ProtobufBodyContent {
    constructor(endpoints: Endpoint[], sn_list: DeviceId[], passive_pn_list: DeviceId[], name?: string);
    endpoints(): Endpoint[];
    sn_list(): DeviceId[];
    passive_pn_list(): DeviceId[];
    name(): string | undefined;
    set_name(name?: string): void;
    try_to_proto(): BuckyResult<protos.DeviceBodyContent>;
}
// ...
export class DeviceBuilder extends NamedObjectBuilder<DeviceDescContent, DeviceBodyContent> {
}
export class DeviceId extends NamedObjectId<DeviceDescContent, DeviceBodyContent> {
    constructor(id: ObjectId);
    static default(): DeviceId;
    static from_base_58(s: string): BuckyResult<DeviceId>;
    static try_from_object_id(id: ObjectId): BuckyResult<DeviceId>;
}
export class DeviceIdDecoder extends NamedObjectIdDecoder<DeviceDescContent, DeviceBodyContent> {
    constructor();
}
export class Device extends NamedObject<DeviceDescContent, DeviceBodyContent> {
    static create(owner: Option<ObjectId>, unique_id: UniqueId, endpoints: Endpoint[], sn_list: DeviceId[], passive_sn_list: DeviceId[], public_key: PublicKey, area: Area, category: DeviceCategory, build?: (builder: DeviceBuilder) => void): Device;
    device_id(): DeviceId;
    connect_info(): DeviceBodyContent;
    name(): string | undefined;
    set_name(name?: string): void;
    category(): BuckyResult<DeviceCategory>;
}
export class DeviceDecoder extends NamedObjectDecoder<DeviceDescContent, DeviceBodyContent, Device> {
    constructor();
}
// ...
export class NamedObjectId<DC extends DescContent, BC extends BodyContent> implements RawEncode, Compareable<NamedObjectId<DC, BC>> {
        constructor(obj_type: number, object_id: ObjectId);
        get object_id(): ObjectId;
        hashCode(): symbol;
        equals<ODC extends DescContent, OBC extends BodyContent>(other: NamedObjectId<ODC, OBC>): boolean;
        into(): ObjectId;
        gen(object_id: ObjectId): void;
        toString(): string;
        toJSON(): string;
        to_string(): string;
        to_base_58(): string;
        raw_measure(ctx?: any): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
// ...
export class NamedObject<DC extends DescContent, BC extends BodyContent> implements RawEncode {
        constructor(desc: NamedObjectDesc<DC>, body: Option<ObjectMutBody<DC, BC>>, signs: ObjectSigns, nonce: Option<JSBI>);
        obj_type(): number;
        obj_type_code(): number;
        calculate_id(): ObjectId;
        protected obj_flags(): number;
        to_vec(): BuckyResult<Uint8Array>;
        to_hex(): BuckyResult<string>;
        toString(): string;
        to_string(): string;
        toJSON(): string;
        desc(): NamedObjectDesc<DC>;
        body(): Option<ObjectMutBody<DC, BC>>;
        set_body(body: Option<ObjectMutBody<DC, BC>>): void;
        body_expect(): ObjectMutBody<DC, BC>;
        signs(): ObjectSigns;
        nonce(): Option<JSBI>;
        raw_measure(_ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: NamedObjectContext, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf(purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        raw_hash_encode(): BuckyResult<HashValue>;
}
// ...
export enum ObjectCategory {
    Standard = "standard",
    Core = "core",
    DecApp = "dec_app"
}
// ...
export type ObjectIdInfo = (StandardObjectIdInfo | CoreObjectIdInfo | DecAppObjectIdInfo) & ObjectIdInfoMatcher;
export class ObjectId implements RawEncode, Compareable<ObjectId> {
    m_buf: Uint8Array;
    m_base58?: string;
    constructor(buf: Uint8Array);
    get object_id(): void;
    static default(): ObjectId;
    static copy_from_slice(buf: Uint8Array): ObjectId;
    obj_type_code(): ObjectTypeCode;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    as_slice(): Uint8Array;
    clone(): ObjectId;
    length(): number;
    toString(): string;
    toJSON(): string;
    to_string(): string;
    to_base_58(): string;
    hashCode(): symbol;
    equals(other: ObjectId): boolean;
    eq(other: ObjectId): boolean;
    static from_str(s: string): BuckyResult<ObjectId>;
    static from_base_58(s: string): BuckyResult<ObjectId>;
    to_hash_value(): HashValue;
    object_category(): ObjectCategory;
    is_stand_object(): boolean;
    is_core_object(): boolean;
    is_dec_app_object(): boolean;
    info(): ObjectIdInfo;
}
// ...
export enum ObjectTypeCode {
    Device = 1,
    People = 2,
    SimpleGroup = 3,
    Org = 4,
    AppGroup = 5,
    UnionAccount = 6,
    Chunk = 7,
    File = 8,
    Dir = 9,
    Diff = 10,
    ProofOfService = 11,
    Tx = 12,
    Action = 13,
    ObjectMap = 14,
    Contract = 15,
    Custom = 16
}
// ...
export class UniqueId implements RawEncode {
    m_buf: Uint8Array;
    constructor(buf: Uint8Array);
    as_slice(): Uint8Array;
    length(): number;
    toJSON(): string;
    static default(): UniqueId;
    static copy_from_slice(buf: Uint8Array): UniqueId;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    to_base_58(): string;
    static from_base_58(s: string): BuckyResult<UniqueId>;
}
// ...
export class DeviceBodyContent extends jspb.Message {
    clearEndpointsList(): void;
    getEndpointsList(): Array<Uint8Array | string>;
    getEndpointsList_asU8(): Array<Uint8Array>;
    getEndpointsList_asB64(): Array<string>;
    setEndpointsList(value: Array<Uint8Array | string>): void;
    addEndpoints(value: Uint8Array | string, index?: number): Uint8Array | string;

    clearSnListList(): void;
    getSnListList(): Array<Uint8Array | string>;
    getSnListList_asU8(): Array<Uint8Array>;
    getSnListList_asB64(): Array<string>;
    setSnListList(value: Array<Uint8Array | string>): void;
    addSnList(value: Uint8Array | string, index?: number): Uint8Array | string;

    clearPassivePnListList(): void;
    getPassivePnListList(): Array<Uint8Array | string>;
    getPassivePnListList_asU8(): Array<Uint8Array>;
    getPassivePnListList_asB64(): Array<string>;
    setPassivePnListList(value: Array<Uint8Array | string>): void;
    addPassivePnList(value: Uint8Array | string, index?: number): Uint8Array | string;

    hasName(): boolean;
    clearName(): void;
    getName(): string;
    setName(value: string): void;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): DeviceBodyContent.AsObject;
    static toObject(includeInstance: boolean, msg: DeviceBodyContent): DeviceBodyContent.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: DeviceBodyContent, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): DeviceBodyContent;
    static deserializeBinaryFromReader(message: DeviceBodyContent, reader: jspb.BinaryReader): DeviceBodyContent;
}